Remove `buffer-alloc` and `buffer-from` deps
`Buffer.alloc` and `Buffer.from` were added in Node 6.

So, this change only drops support for Node 4 and lower, which no one should be running at this point. Especially since Node 6 is already EOL and Node 8 is about to be EOL in December 2019. :)

I also removed package-lock.json which is not useful in packages since it doesn't get published to npm. It's existence only ensures that package authors get outdated deps that the users are not going to get.